+++ src/main/java/ro/upb/smartfeedback/repository/IntrebariRepository.java
@@ -8,5 +8,5 @@ import ro.upb.smartfeedback.entity.Intrebari;
  * Created by colez on 29/11/2015.
  */
 @Repository
-public interface IntrebariRepository extends JpaRepository<Intrebari, Integer>{
+public interface IntrebariRepository extends JpaRepository<Intrebari, Long>{
 }
