+++ src/main/java/ro/upb/smartfeedback/repository/TipActivitateRepository.java
@@ -9,6 +9,6 @@ import ro.upb.smartfeedback.entity.TipActivitate;
 import java.util.List;
 
 @Repository
-public interface TipActivitateRepository extends JpaRepository<TipActivitate, Integer> {
+public interface TipActivitateRepository extends JpaRepository<TipActivitate, Long> {
     public TipActivitate getByNume(String nume);
 }
